[BE] refactor: 자료형 수정
- price 를 Double 에서 Long 으로 변

#30

+++ BE/airbnb/src/main/java/com/codesquad/airbnb/accmmodation/data/Price.java
@@ -12,10 +12,10 @@ import lombok.NoArgsConstructor;
 public class Price {
 
   @Column(nullable = false)
-  private Double price;
+  private Long price;
 
   @Builder
-  public Price(Double price) {
+  public Price(Long price) {
     this.price = price;
   }
 }
